fix: widen frameDataSize from uint8_t to uint16_t
The frame data size is a 16-bit value assembled from two bytes
(frameBuffer[3] << 8 | frameBuffer[4]), but was stored in a uint8_t,
truncating the high byte. This made the FRAME_BUFFER_SIZE guard
unreachable (uint8_t max 255 < 512) and caused incorrect frame
length matching for payloads > 255 bytes.

+++ src/SeeedmmWave.cpp
@@ -349,7 +349,7 @@ void SeeedmmWave::fetch(uint32_t timeout) {
   static bool startFrame = false;
   static std::vector<uint8_t> frameBuffer;
   uint32_t expire_time = millis() + timeout;
-  uint8_t frameDataSize;
+  uint16_t frameDataSize;
   do {
     size_t c_available = _serial->available();
     while (c_available--) {
